Changed add session modal
