+++ apps/web-app/components/EditSessionModal/index.tsx
@@ -8,7 +8,7 @@ import ModalSteps from "./ModalSteps"
 import Step1 from "./Step1"
 import Step2 from "./Step2"
 import { EventsDTO, SessionsDTO } from "../../types"
-import { displayDateWithoutTimezone } from "../../data/dateFormat"
+import { displayDateWithoutTimezone, to24HourFormat, to12HourFormat } from "../../data/dateFormat"
 
 type NewSessionState = {
     description: string
@@ -45,15 +45,6 @@ type Props = {
 }
 
 const EditSessionModal = ({ isOpen, closeModal, session, sessions, events }: Props) => {
-    const to12HourFormat = (time: string) => {
-        const [hour, minute] = time.split(":")
-        const hourInt = parseInt(hour)
-        const amPm = hourInt >= 12 ? "PM" : "AM"
-        const hour12 = hourInt % 12 === 0 ? 12 : hourInt % 12
-        const hourString = hour12 < 10 ? `0${hour12}` : hour12
-        return `${hourString}:${minute} ${amPm}`
-    }
-
     const router = useRouter()
     const questionTextRef = useRef(null)
     const [isLoading, setIsLoading] = useState(false)
@@ -80,21 +71,6 @@ const EditSessionModal = ({ isOpen, closeModal, session, sessions, events }: Pro
         event_slug: session.event_slug,
         event_item_id: session.event_item_id
     })
-
-    const to24HourFormat = (time: string) => {
-        const [timePart, amPm] = time.split(" ")
-        const [hour, minute] = timePart.split(":")
-
-        let hourInt = parseInt(hour)
-        if (amPm === "PM" && hourInt !== 12) {
-            hourInt += 12
-        } else if (amPm === "AM" && hourInt === 12) {
-            hourInt = 0
-        }
-
-        const hourString = hourInt < 10 ? `0${hourInt}` : hourInt
-        return `${hourString}:${minute}`
-    }
 
     const handleSubmit = async () => {
         setIsLoading(true)
