+++ apps/web-app/package.json
@@ -18,6 +18,8 @@
         "@tailwindcss/line-clamp": "^0.4.2",
         "autoprefixer": "^10.4.14",
         "axios": "^0.27.2",
+        "date-fns": "^2.29.3",
+        "date-fns-tz": "^2.0.0",
         "dotenv": "^16.0.1",
         "draft-js": "^0.11.7",
         "draft-js-export-html": "^1.4.1",
