+++ apps/web-app/pages/api/fetchLocations.tsx
@@ -7,7 +7,6 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
     const supabase = createServerSupabaseClient({ req, res })
     try {
         const response = await supabase.from("locations").select("*")
-        console.log("RESPONSE LOCATIONS", response)
         if (response.error === null) res.status(200).send(response.data)
         else res.status(response.status).send(response.error)
     } catch (err: any) {
